TETO-102 Refactor: feign를 프로젝트 내 코드로 리팩토링
