+++ backend/restaurant-service/src/main/java/com/tabletopia/restaurantservice/domain/reservation/dto/TableSnapshot.java
@@ -1,4 +1,4 @@
-package com.tabletopia.restaurantservice.dto;
+package com.tabletopia.restaurantservice.domain.reservation.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
